Update Progress Day #2
I've commented a lot of the functionality and the peeves I have with my lack of abilities. So, far I have no issue except for the lack of optimization / better way to do it.

Furthermore, the bot is almost ready in a way, having the core commands prepared -- Search, Reply and Character Info.
Soon, I'll be adding cogs to place all of them there and calculator functionality.
Eh... I guess that's all for now. Bed time.